Remove unrelated utility
