+++ src/sudoku/exp2.js
@@ -1,10 +1,6 @@
 import { range, shuffle } from '../util/arrays.js';
-// import Debugger from '../util/debug.js';
 import Sudoku, { cellMask, cellsFromMask } from './Sudoku.js';
 import SudokuSieve from './SudokuSieve.js';
-
-// const isDebugging = true;
-// const whenDebugging = new Debugger(isDebugging);
 
 /**
  * @typedef {object} Node
