+++ src/sudoku/sudoku-puzzle-driver.js
@@ -1,21 +1,18 @@
-import Debugger from '../util/debug.js';
 import Sudoku from './Sudoku.js';
-
-const debug = new Debugger(true);
 
 const resultsForNonSieve = Sudoku.generate({
   amount: 100,
   numClues: 24,
   useSieve: true,
-  callback: (generated) => { debug.log('+ ', generated.toString()); }
+  callback: (generated) => { console.log('+ ', generated.toString()); }
 });
 
-debug.log('Generated:', resultsForNonSieve.length);
+console.log('Generated:', resultsForNonSieve.length);
 const timesNonSieve = [];
 resultsForNonSieve.forEach((result, index) => {
   const { puzzle, pops, resets, timeMs } = result;
   timesNonSieve.push(timeMs);
-  debug.log(`${puzzle.toString()}\n    { time: ${timeMs}ms, pops: ${pops}, resets: ${resets} }`);
+  console.log(`${puzzle.toString()}\n    { time: ${timeMs}ms, pops: ${pops}, resets: ${resets} }`);
 });
 
 // Print out the total mean time, interquartile mean time, and median time, and standard deviation
@@ -36,4 +33,4 @@ const iqrMean = sorted.reduce((acc, time) => {
 const median = sorted[Math.floor(timesNonSieve.length / 2)];
 const variance = timesNonSieve.reduce((acc, time) => acc + Math.pow(time - mean, 2), 0) / timesNonSieve.length;
 const stdDev = Math.sqrt(variance);
-debug.log(`\nMean: ${mean}ms, IQR Mean: ${iqrMean}ms, Median: ${median}ms, Standard Deviation: ${stdDev}ms`);
+console.log(`\nMean: ${mean}ms, IQR Mean: ${iqrMean}ms, Median: ${median}ms, Standard Deviation: ${stdDev}ms`);
