--- src/util/debug.js
@@ -1,11 +1,0 @@
-export default class Debugger {
-  constructor(enabled = true) {
-    this._enabled = Boolean(enabled)
-  }
-
-  log(...things) {
-    if (this._enabled) {
-      console.log(...things);
-    }
-  }
-}
